+++ src/features/post/index.ts
@@ -1,1 +1,2 @@
 export * from "./model";
+export * from "./ui";
