+++ src/features/post/model/usePostStore.ts
@@ -6,6 +6,7 @@ interface PostState {
   total: number;
   selectedPost: PostWithUser | null;
   newPost: NewPost;
+  setPosts: (posts: PostWithUser[]) => void;
   addPost: (post: PostWithUser) => void;
   updatePost: (updatedPost: PostWithUser) => void;
   deletePost: (postId: number) => void;
@@ -18,6 +19,7 @@ export const usePostStore = create<PostState>((set) => ({
   total: 0,
   selectedPost: null,
   newPost: { title: "", body: "", userId: 1 },
+  setPosts: (posts) => set({ posts, total: posts.length }),
   addPost: (post) => set((state) => ({ posts: [...state.posts, post] })),
   updatePost: (updatedPost) =>
     set((state) => ({ posts: state.posts.map((post) => (post.id === updatedPost.id ? updatedPost : post)) })),
