+++ src/features/post/ui/PostTableRow.tsx
@@ -1,0 +1,105 @@
+import { deletePost, Post, PostWithUser } from "@entities/post";
+import { Button, TableCell, TableRow } from "@shared/ui";
+import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react";
+import { usePostQuery, usePostStore } from "../model";
+import { getUser, User } from "@entities/user";
+
+const PostTableRow = (post: PostWithUser) => {
+  const { posts, setPosts, setSelectedPost } = usePostStore();
+  const { searchQuery, updatePostSearchParams } = usePostQuery();
+
+  const fetchPostsByTag = async (tag: string) => {
+    if (tag === "all") {
+      updatePostSearchParams("selectedTag", "");
+      return;
+    }
+    updatePostSearchParams("selectedTag", tag);
+  };
+
+  const deletePostAndUpdate = async (id: number) => {
+    await deletePost(id);
+    setPosts(posts.filter((post) => post.id !== id));
+  };
+
+  // 게시물 상세 보기
+  const openPostDetail = (post: Post) => {
+    setSelectedPost(post);
+    setShowPostDetailDialog(true);
+  };
+
+  // 사용자 모달 열기
+  const openUserModal = async (user: User) => {
+    try {
+      const userData = await getUser(user.id);
+      setSelectedUser(userData);
+      setShowUserModal(true);
+    } catch (error) {
+      console.error("사용자 정보 가져오기 오류:", error);
+    }
+  };
+
+  return (
+    <TableRow key={post.id}>
+      <TableCell>{post.id}</TableCell>
+      <TableCell>
+        <div className="space-y-1">
+          <div>{highlightText(post.title, searchQuery)}</div>
+
+          <div className="flex flex-wrap gap-1">
+            {post.tags?.map((tag) => (
+              <span
+                key={tag}
+                className={`px-1 text-[9px] font-semibold rounded-[4px] cursor-pointer ${
+                  selectedTag === tag
+                    ? "text-white bg-blue-500 hover:bg-blue-600"
+                    : "text-blue-800 bg-blue-100 hover:bg-blue-200"
+                }`}
+                onClick={() => {
+                  fetchPostsByTag(tag);
+                }}
+              >
+                {tag}
+              </span>
+            ))}
+          </div>
+        </div>
+      </TableCell>
+      <TableCell>
+        <div className="flex items-center space-x-2 cursor-pointer" onClick={() => openUserModal(post.author)}>
+          <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
+          <span>{post.author?.username}</span>
+        </div>
+      </TableCell>
+      <TableCell>
+        <div className="flex items-center gap-2">
+          <ThumbsUp className="w-4 h-4" />
+          <span>{post.reactions?.likes || 0}</span>
+          <ThumbsDown className="w-4 h-4" />
+          <span>{post.reactions?.dislikes || 0}</span>
+        </div>
+      </TableCell>
+      <TableCell>
+        <div className="flex items-center gap-2">
+          <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
+            <MessageSquare className="w-4 h-4" />
+          </Button>
+          <Button
+            variant="ghost"
+            size="sm"
+            onClick={() => {
+              setSelectedPost(post);
+              setShowEditDialog(true);
+            }}
+          >
+            <Edit2 className="w-4 h-4" />
+          </Button>
+          <Button variant="ghost" size="sm" onClick={() => deletePostAndUpdate(post.id)}>
+            <Trash2 className="w-4 h-4" />
+          </Button>
+        </div>
+      </TableCell>
+    </TableRow>
+  );
+};
+
+export default PostTableRow;
