+++ src/features/post/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./PostTable";
+export * from "./PostTableRow";
